--- constants.js
@@ -1,3 +1,0 @@
-export const FETCHING_DATA = 'FETCHING_DATA'
-export const FETCHING_DATA_SUCCESS = 'FETCHING_DATA_SUCCESS'
-export const FETCHING_DATA_FAILURE = 'FETCHING_DATA_FAILURE'
