+++ reducers/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux'
+
+import content, * as contentDuck from './content'
+
+export default combineReducers({
+  content
+})
